+++ src/consts.ts
@@ -56,6 +56,11 @@ export const SCREEN_WIDTH = 20;
 export const SCREEN_HEIGHT = 18;
 export const TILE_SIZE = 8;
 
+export const SCREEN_WIDTH_PX = SCREEN_WIDTH * TILE_SIZE;
+export const SCREEN_HEIGHT_PX = SCREEN_HEIGHT * TILE_SIZE;
+
+export const SCENE_MAX_SIZE_PX = 2040;
+
 // Scripts
 export const MAX_NESTED_SCRIPT_DEPTH = 5;